Rename "scans" app to "sites"
This was a fairly straightforward `git mv`, s/scans/sites/g, and then a
little bit of manual cleanup.

The important thing to note is that this change does _not_ attempt to
migrate the database from scans to sites. While that is certainly
possible, it does not seem worth the effort for a site that is still in
the early development phase. However, the old migrations are preserved
so they can be reapplied. After checking this commit out, you should:

1. Destroy your old development database
2. `./manage.py migrate`
3. `./manage.py loaddata` for the newest fixtures in `sites/fixtures`,
   which includes the new slugs for each Site.