Remove BlogPost intro
The intro field was awkward in a lot of ways. It was not included in the
individual blog post views, so there was a weird disconnect between the
summary of each post on the index page and the posts themselves.
Including the intro created other issues, because it was a CharField and
not a RichTextField, so that created a weird limitation on styles in the
first paragraph or so of an article. I also ran into annoying issues
with the Sass that I didn't feel like spending time debugging.

The intro field was also unnecessary, because it is just as easy to
extract the initial text of the blog post's body to use for the summary
on the index page using the builtin `truncatewords_html` template tag.
Therefore, I think the simplest and best solution is to remove it.