Only allow one HomePage to be created
Setting parent_page_types = [] prevents a Page from being created at
all. According to the docs, "this is useful for unique pages that should
only be created once":

http://docs.wagtail.io/en/v1.6.2/reference/pages/model_reference.html#wagtail.wagtailcore.models.Page.parent_page_types

We only need one HomePage in our design, so applying this restriction
makes sense. It also improves the ergonomics of adding child pages in
the Wagtail Admin because there is no longer an option to add a "Home
Page", which was unnecessary clutter.