Remove "Strictly Forces HTTPS"
After getting clarification from the pshtt devs, "Strictly Forces HTTPS"
is not a useful component for our metric. It is useful for pshtt because
it includes redirects to another (secure) origin, which is a somewhat
common occurrence among .gov domains. That is not a common situation for
most websites, including the type of sites that this project aims to
evaluate.

- Updates the Scan model and adds a migration for the change.
- Updates the scoring function.
- Removes references to "Enforces HTTPS" from the leaderboard.
- Updates the development database fixtures.