+/- for letter grades
DRY out computing the grade and the CSS class name. It feels wrong to
compute view logic (like the CSS class name) in the model, but this
approach allows us to avoid duplicating the same logic for computing the
grade and class name in both JS (for the leaderboard) and Python (for
the site score breakdown pages).

STN-44 #done