Move floating table header into Backbone View
Refactors the floating table header code so it is self-contained within
the Backbone View, per Gabe's review comments on #28.

This commit implements the floating table header without using a plugin.
Since we are no longer using floatThead, I removed it from package.json.

Removes some styles from _leaderboard.scss that did not appear to be
neccessary (I could not visually detect any difference after removing
them) and that were also interfering with the layout of the floating
table header. In particular, `table { width: 100% }`, because it means
different things in the context of the fixed position floating table
header vs. the relatively positioned normal table header.

Tested successfully on the latest Chrome, Firefox, and Safari on Mac OS
10.10. In fact, this custom solution works better in Safari than
floatThead did (it was a little glitchy during the transition from
floating to fixed header because some of the table sizing would change
slightly).