Implement site score teaser
Adds a Backbone View for the site score teaser on the homepage.

It uses the same basic approach as the Leaderboard for instantiating
itself with JSON injected into the template. This approach is superior
to randomizing the sites on the server, because it improves the
cacheability of the homepage (at the cost of increasing the size of the
page, but it's not by much).