Add "Top 10" leaderboard to homepage
Modifies the Leaderboard view so the constructor takes additional
options. This allows us to reuse the same code for both the homepage's
"top 10" leaderboard and the full leaderboard.

Options are extracted from data attributes on the View element. One
approach would've been to do this extraction within the View's
`initialize` method (e.g. via `.$el.data()`), but I decided it was
better to decouple the view options from the underlying DOM Element..
This makes it easy to use data attributes to set these options, but
allows callers to set them explicitly in the call to the constructor if
desired.

Note that options are stored separately from the view state. This is
because they are set on construction and are expected to remain fixed
for the lifetime of the view.