Update Python dependencies and pin wagtailmenus
Updates Django, Wagtail, and Pillow to the latest versions so we can
take advantage of various improvements in recent releases.

Updates wagtailmenus to the latest version. Initially, this caused an
AttributeError on every page load. I fixed this by adding the
wagtailmenus context processor, which became a requirement in
wagtailmenus==1.6.0.

Pins wagtailmenus to the latest version. This means all of our Python
dependencies are now pinned, which should avoid breaking changes due to
surprise dependency updates in the future.