Update instructions for developers in README
Testing the update of the Python dependencies provided me with an
opportunity to verify the "Getting Started" instructions for developers
in the README. I verified that following the instructions to set up a
new Secure the News developer environment from scratch works as
expected.

Update the documentation to explain the new tmux environment for
developers, and to clarify first steps after setting up a new
development environment (run migrations and load the development
fixtures).

Fix incorrect version requirements for developer prerequisites.