Basic moderation for Pledges via wagtailmodeladmin
Adds an entry in the sidebar to view and edit Pledges through the
Wagtail admin. This enables a basic approval workflow:

1. Check Pledges in Wagtail admin
2. Review Pledges that have not yet been approved
3. If a pledge looks good, approve it and save the change to the model.
4. Otherwise, delete it.

Ideally, the only field of the Pledge model that would be editable in
the admin would be the "Approved" BooleanField. Making the other fields
read-only would remove the potential for an admin to accidentally modify
a pledge. Unfortunately, this is not currently possible in
wagtailmodeladmin without significant customization. There is an open
issue: https://github.com/torchbox/wagtail/issues/2893.