Pledge status
Replace Pledge.approved, which was a BooleanField with only two choices,
with Pledge.status, which supports 3 possible status: "Needs Review",
"Approved", and "Rejected".

This makes the Pledge moderation workflow more explicit, allows us to
keep a historical record of all submitted pledges (previously, the
workflow implied that a pledge should be deleted if it were not
approved), and allows us to provide a more helpful pledge status page to
end users.