Update admin user's fixture
Django 1.10 increased the default iteration count for PBKDF2:
https://docs.djangoproject.com/en/1.10/releases/1.10/#django-contrib-auth.

If we don't update this in the fixtures, Django will auto-update it
every time, which will clutter future fixtures exports. Better to take
care of it now.

Note that this "admin" user is only intended for development and should
not be included in a production deployment's database - their
credentials are exposed because the password hash is included in the
development fixtures (and the plaintext password is included in the
README).