Updates prod settings for staticfiles locations
The "base" settings correctly specified two locations for staticfiles,
one for gulp/nodejs and the other for django's staticassets. The prod
settings were inappropriately overriding this sane setting, and
including only the gulp/nodejs files, but excluding the django
staticfiles.

Fixed by removing the override in the prod environment settings; the
base settings now apply mututally to dev and prod.