Rework pledge submission flow
Previously, pledges were made available for review as soon as they have
been submitted. This commit changes that flow to require that a pledge
is confirmed before it can be reviewed.

When a pledge is submitted, a confirmation email is sent to the provided
email address (which must be on the domain or a subdomain of the
associated Site). The email contains a link with an unguessable nonce,
which ensures that only someone who can receive email at the given
address under the site's domain can confirm a pledge. Clicking the link
opens a form which prompts the user to confirm the submission, after
which it will become available for us to review.

The benefit of this approach is that it provides a higher degree of
verification for the pledge submission. Anybody can submit a pledge, but
only someone who is able to receive email under a news organization's
domain name can confirm a submission. This should help reduce noise for
admins who are moderating pledge submissions.

Obviously email is not perfectly secure, but this requirements raises
the stakes significantly and is suitable for this low-risk application.
This is basically the same technique used for password reset emails (or
for verifying DV HTTPS certificates - yikes!).

One potential issue that I can forsee with this approach is that, since
emails are automatically sent to the provided contact email address, it
could be used to spam employees of news organizations with confusing
emails. The current email copy explains that nothing will happen if they
choose not to confirm the submission, and encourages them to contact us
if they have concerns. We should also consider implementing rate
limiting to minimize the impact of anybody who wants to use this form to
spam news organization employees.