Refactor email sending code into views.py
It felt gross to have this code in models.py, so I refactored it into
the views (conceptually, emails are views).

The only challenge was sending the review notification email, since that
is triggered by saving a Pledge with STATUS_APPROVED, which is handled
by wagtailmodeladmin and not one of the pledges views. Originally I was
overriding Pledge.save, but it felt weird to import code from the views
into the models (and created a risk of circular dependencies). I decided
that, while Django's signals are generally recommended as a last resort,
that this was a perfect use case for them - performing an action
triggered by code outside of your control. Works great!