Unicode-aware slugs for Sites
Django's built-in slugify discards characters that cannot be converted
to ASCII. In at least one case this caused a site which had a name that
was entirely non-ASCII characters to end up with an empty slug, which is
obviously broken.

URLs are limited to ASCII characters only, but you can percent-encode
non-ASCII characters and modern browsers are smart enough to display
them correctly and handle them usefully (e.g. in the context of
copy-paste). Django is also smart enough to handle automatically
encoding URLs with non-ASCII characters *as long as you use `reverse`
and friends*. For details, see
https://docs.djangoproject.com/en/1.10/ref/unicode/#taking-care-in-get-absolute-url.