Only display results from scanned sites
If a new Site is added (e.g. via the 'News Sites' wagtailmodeladmin
interface), it will not have any scan data available until the next time
a scan is run (either automatically or manually). A lot of the view code
assumes that all Sites have at least one Scan, so prior to this commit
adding a Site without any Scans caused massive breakage.

Fortunately, thanks to Django's custom Model Managers, this was
extremely easy to fix in a clear, concise, and DRY manner!