Quick fix for broken Vagrant builds
1ec6be01 added psycopg2 to requirements.txt, but did not add the
system-level package manager dependencies required for a successful `pip
install psycopg2`. This is because the Vagrant development environment
and the WIP production deployment mechanisms and quite different. We're
working on combining them in a more natural way, but for now this is a
quick fix that allows `vagrant up` to work again.