Add meta tags for Twitter Cards
Reasonable default metadata is defined in base.html, which is the base
template for every page on Secure the News, so every page on Secure the
News will display a Twitter Card. In addition, Twitter cards for Site
pages and BlogPosts have their title and description tailored to be
relevant to the specific Site or BlogPost being shared.