Purge frontend cache when saving Scans
Whenever a Scan is saved, purges the corresponding site-specific page,
the leaderboard, and the homepage. This automatically keeps the scores
visible on the site up-to-date with the latest scan data.

Leverages Wagtail's frontend cache invalidator, and its
purge_url_from_cache utility function, to do frontend cache invalidation
for pages even though they are not Wagtail pages.

In the common "mass scan" use case (`python manage.py scan`, which is
run periodically by a cron job), this approach is slightly inefficient:
it sends O(n) cache invalidation requests for the leaderboard and
homepage, one for each site, in rapid succession. It would be nice to
handle the "mass scan" use case more efficiently, but since the number
of sites is relatively low (~100), I think the current approach is
sufficient for now.