Pin pshtt to version that fixes downgrades_https bug
I wrote an upstream fix for the downgrades_https null bug, which was merged in
https://github.com/dhs-ncats/pshtt/pull/50. Since pshtt hasn't cut a new
release, let's pin to the commit with the fix on the upstream repo for now.

Note that this change only affects the development environment. An equivalent
change will need to be merged into the infrastructure repo for this fix to
reach production.