Optimize data migration
Since this is such a simple change, we can optimize it by using a bulk update
call instead of iterating over all of the Scans in the database. I expect this
to be helpful on production systems which have many historical scans stored
that will be updated by this data migration.

Note that bulk updates must be used carefully, because they do not call custom
save methods or fire any save-related signals. While Scans do have a custom
save method, we do not need to call it again for this data migration because
the code in `_score` uses `if downgrades_https`, so None is equivalent to False
and the scores would not change anyway.