Pin django-analytical to fix mixed content issue
django-analytical's Piwik integration introduced a mixed content issue on
Secure the News for users with Javascript disabled because it hard-coded an
http scheme for the fallback tracking pixel. I fixed the issue upstream in
https://github.com/jcassee/django-analytical/pull/101, but the maintainers
haven't cut a new release yet, so for now we will pin to the fixed commit on
the upstream master branch.