Clarify choosing between Vagrant providers
According to the Vagrant docs, the default provider is `virtualbox`;
however, through testing (on Vagrant 1.8.5), I found that the first
provider listed in the Vagrantfile is used by default.

Rearranges the Vagrantfile so the `virtualbox` provider is listed first
and therefore used by default (satisfying the principle of least
surprise). Clarifies the development environment setup docs to explain
that you only need to set `VAGRANT_DEFAULT_PROVIDER` if you wish to use
a non-default provider.