Order posts on blog index page by primary key
Previously, blog posts on the index page were ordered by the
date column. If two blog posts were posted on the same day,
the most recent blog post may not appear at the top of the index
page. This was bug #97 on the GitHub issue tracker.

This commit fixes this by ordering the index page by the
primary key (order by desc) instead of the date column.