Builds static assets via Gulp in CI flow
Left out the gulp logic when porting from Vagrant to Molecule.
Readds the `gulp build:production` command necessary for `./manage.py
collectstatic` to work. Required overriding the NodeJS build container,
which is now supported upstream in the django-stack role.

Force nodejs v4.8.2, not 6.11.0, since that's what the application was
built and tested against, and also what it still runs in production
today.

Note that we're not pruning the "src" dir from the rsync command, since
that also included the `src/client` directory, which is required for the
gulp build commands to succeed. (The gulp build commands write to
`src/build`, and the Django collectstatic action looks there.)