Switch node's gulp to utilize watch target
This is a better setting for developers so that there is an active listener
against node assets that are actively refreshed. Had to tweak the logic a bit so
that ansible waited until the container was done installing before proceeding
since the container will always be running when the watch logic is hit.