Patches SlugField to accept Unicode
We were already explicitly allowing Unicode in the `slugify` call that
generates the slug for a news site prior to storing on the model, but
the model's SlugField *also* needs to be told explicitly that it may
store Unicode values. See docs [0].

Includes corresponding database migration, so that the prod db is up to
speed on the next deploy.

Includes a new app test to ensure that the Unicode slug handling doesn't
break on model validation during save. Will need to be updated once #127
is resolved.

Closes #130.

[0] https://docs.djangoproject.com/en/1.11/ref/models/fields/#django.db.models.SlugField.allow_unicode