Adds outdated "home" app migration
The "migrate" command was complaining that not all migrations had been
applied, so generated this one. It should have been done a while ago,
but the change is minor, and it doesn't appear to have been used in the
lag time between the code changing and the db not catching up.