Add SiteCategory and enable admins to add/edit via Wagtail
Adds generic categories that administrators can add via Wagtail by
navigating to Snippets -> Site Categories.

The category has a name (string to be displayed on the
leaderboard) along with an optional image that also can be
uploaded via Wagtail. The motivation here was in case we want to
use cute flags like we do for the SecureDrop interface language
picker.