Convert developer env to docker-compose
In addition to what the title says, performed the following:

* Bumped python from 3.4 --> 3.6
* Scraped molecule dev scenario
* Converted some of the one-off docker exec commands to compose compat.
* Added in pipenv file for the developer scenario
* Pegged dependencies for python container

This last bit is important, since this container will potentially be
used in production, its important we have historical record, stability,
introduced into the dependencies here. This means we'll need to utilize
some kind of vulnerability scanner to determine when to update tho.
Quay.io does this, we can also examine others in CI.