Provide scaffolding to build prod django container
Want to add a little abstraction in a later commit to keep things DRY in
the dev vs. prod environment.

I'm trying to make abstract the docker-compose file here so we can
over-ride env. vars for use in production. That may require some more
ironing out here but I think this is a good start.