Refactors container entrypoint script
Ported ash -> bash, for better error handling. Removed dependency on the-
"wait-for-it" script, opting instead to use `nc -z` directly. The sole-
loss of functionality here is that the `nc -z` will retry indefinitely.
We can add logic to timeout if that becomes a problem.

Also inverts the conditional logic governing prod/dev functionality.
Now, the script defaults to prod functionality (e.g. prefer gunicorn
over runserver), running the dev-only logic if-and-only-if
DEPLOY_ENV=dev. Hopefully this prevents misconfigurations in prod.