Add image name to prod docker-compose
This changes we no longer have to do tag wrangling post build. It does
mean that if someone were to use the docker-compose file directly they
might be running against a prod image rather than a local image. Thats
always a problem though and the makefile cmds ensure we force a rebuild.