Ensures .env file populated for prod container build
The `.env` file is required, but wasn't being created via the prod
container build flow. It'll be in the local dir if the dev scenario has
already been run, but let's set a dependency in the Makefile target to
be sure. Uncovered this by running `git clean -f -x` to remove the
concatenated dockerfiles.