Ignore CVE-2018-18074
There is a CVE associated to `requests` library:
https://nvd.nist.gov/vuln/detail/CVE-2018-18074

The vulnerability is that HTTP authorization headers are not stripped
when redirected to HTTP from HTTPS. Authorization Headers are not used
by the scanner, furthemore the scanner (and therefore `requests` only
scans public endpoints.

`requests`, `pyopenssl` and `pshht` libraries are very tightly coupled,
and there is a hard requirement for Python 3.6 as part of upgrade of
for `pyopenssl`. The upgrade is tracked in #174.