Ignore pyopenssl DoS vulnerabilities
`pyopenssl` < 17.5.0 denial of service vulnerability while parsing and
storing x509 certificates.

Since we control the endpoints that are being scanned, a potential
adversary cannot trigger these DoS vulnerabilities without modifying a
certificate on a new organization page, and this change
would also be easy to revert on our side.

We should obviously upgrade these libraries in the future, but it is
currently blocked on pshtt requiring Python >=3.6. This can be tracked
in #174