Strip hard-python requirements
These requirements are problematic because there is no concept in pipenv
of "any version of python 3" that works. Specifically, it'll break usage
of `pipenv check`. So the work-around here is it's up to developers to
specify the python version to use when they initiate the environment. I
don't see this as a huge problem, these are all developer dependencies
here and don't affect server-side