Unbreaks node container build process
There are a few small changes to the container configs here, mostly
obvious cleanup such as removing duplicate lines, or renaming the volume
mounts to match what we use in other django website repos, simply for
consistency.

The major change here is the use of an env var to customize the
webpacker watcher logic. Without that env var being set, webpack hung,
and the node deps never came up on time.