Implement hash pinning for pip requirements
These changes are based on similar ones done to the freedom.press
repository, see

freedomofpress/freedom.press#671

For full context.

In addition, I have taken this opportunity to move the Makefile based
interface for updating the *requirements.in files to match that of our
freedom.press situation.  Specifically, we now have three commands:

* `make compile-pip-dependencies` -- updates all requirements.txt files
based on the content of the requirements.in files
* `make pip-update` -- updates a specific package appearing in requirements.in
* `make pip-dev-update` -- updates a specific package appearing in dev-requirements.in