Uses exec in container entrypoint
The container image should report its process as `gunicorn` in prod
contexts, so that the application gracefully terminates when it receives
a SIGTERM from the managing process [0]. Without using `exec`, the container
reports that sh is the active process (sh -> bash -> gunicorn), so a
SIGTERM to sh results in SIGKILL on the children.

Technically we're customizing the "CMD", not using "ENTRYPOINT"; we
leave the default ENTRYPOINT of "/bin/sh -c" mostly to support ad-hoc
commands in CI/testing, so we don't have to pass `--entrypoint` to each
docker command.

[0] https://docs.gunicorn.org/en/stable/signals.html