Update requirements for python 3.7
This change is needed before we can install software on the containers
running python 3.7

I'm not exactly sure why libpq-dev was needed, but the pip-tools
commands wouldn't work until I installed it.  It was saying that
`pg_config` was needed.  Something to do with psycopg2?

It's also possible there are version problems with pshtt (again).  The
latest version, 0.6.7, will not compile under pip-tools because it has
conflicting requirements with itself:

```
There are incompatible versions in the resolved dependencies:
  cryptography>=2.8 (from pyopenssl==19.1.0->pshtt==0.6.7->-r requirements.in (line 13))
  cryptography==2.5 (from sslyze==2.1.4->pshtt==0.6.7->-r requirements.in (line 13))
```

So I installed the most recent version that successfully compiles.
It's possible there are bugs?  See
https://github.com/cisagov/pshtt/pull/211 for more information.