app: scan for onion service support on news organization sites
Onion-Location is a header for advertising you have an onion service [0].
Onion services protect reader privacy, and enable folks in censored regions
to reach censored news organizations. Tor Browser uses the Onion-Location header
in order to display a "onion available" icon for users to click on.

The Onion-Location (like other HTTP headers) is not case sensitive, hence the
.lower() before checking for the presence of the onion-location key in the headers
dict.

Also note that "onion available" is only shown in Tor Browser [1] and the redirect
to onion only occurs  [2] if the site is over HTTPS.

[0] https://gitweb.torproject.org/tor-browser-spec.git/tree/proposals/100-onion-location-header.txt
[1] https://github.com/acatarineu/tor-browser/commit/0721f13faa499e80c4c15b06444897ab1badf51c#diff-84dbdcf5e7c0177889f69f8e36ec8ea1R47
[2] https://github.com/acatarineu/tor-browser/commit/3a9929a559c8505f3032f8f317fbb600d1a71e92#diff-70a990c1d10c050c5fcc69b226c33c5eR2647-R2648