Rename `onion_available` scan field to `onion_location_header`
Renaming this field better clarify that it refers to the presence or
absence of the `Onion-Location` header (or its meta tag equivalent) in
the site at the time it was scanned.

We are doing this in preparation for incorporating a field on `Site`
that permits overriding this field on a per-site basis.