Update other npm packages.
This changeset was generated with the `npm update` command.

This was the simplest thing I could do that fixed the failures on `npm
audit`, and I think it's okay (possibly even good?) to do this
occasionally to get our JS packages reasonably up-to-date.  The
updates are mostly minor or bugfix versions, and the dev site works
exactly the same as it did before with respect to client side code and
the dev workflow in general.