Ignore safety check for cryptography version <3.3
We can't upgrade because pshtt depends on an old version of sslyze
which does not support the latest cryptography.

This is the information from safety:

39252: cryptography <3.3 resolved (3.2.1 installed)!
Cryptography 3.3 no longer allows loading of finite field Diffie-Hellman parameters of less than 512 bits in length. This change is to conform with an upcoming OpenSSL release that no longer supports smaller sizes. These keys were already wildly insecure and should not have been used in any application outside of testing.