Remove safety-check related tools from pipenv
This change removes `flake8`, `safety` and `bandit` from Pipenv and
adjusts the makefile (and related scripts) to run them either inside a
docker container or directly in the CI environment.

The reason for doing this is a recent update to Pipenv seemed to be
breaking the make commands.  Not fully sure what was going on with it
but it seems like we want to move away from Pipenv anyway and it's not
too hard to remove these packages from it.